Issue #12: Refactoring
