--- web/src/components/Header/Menu/General/index.js
@@ -1,1 +1,0 @@
-export { default } from './MenuGeneral';
