--- web/src/components/Header/Menu/Settings/MenuSettings-styles.js
@@ -1,11 +1,0 @@
-import { css } from '@emotion/react';
-
-const useStyles = () => {
-  return {
-    list: css`
-      width: 240px;
-    `,
-  };
-};
-
-export default useStyles;
