--- web/src/components/Header/Menu/Settings/index.js
@@ -1,1 +1,0 @@
-export { default } from "./MenuSettings";
