--- web/src/components/Header/Menu/index.js
@@ -1,1 +1,0 @@
-export { default } from "./HeaderMenu";
